feat: get new base class
